Add Scoruby namespace and fix tests
