--- lib/decision.rb
@@ -1,15 +1,0 @@
-class Decision
-
-  attr_reader :score, :score_distribution
-
-  def initialize(score, score_distributions)
-    @score = score
-    return if score_distributions.empty?
-
-    @score_distribution = {}
-    score_distributions.each { |score_distribution|
-      attributes = score_distribution.attributes
-      @score_distribution[attributes['value'].to_s]  = attributes['probability'].to_s
-    }
-  end
-end
