--- lib/predicates/false_predicate.rb
@@ -1,13 +1,0 @@
-class FalsePredicate
-  def field
-    nil
-  end
-
-  def true?(_)
-    false
-  end
-
-  def is_missing?(_)
-    false
-  end
-end
