--- lib/predicates/true_predicate.rb
@@ -1,13 +1,0 @@
-class TruePredicate
-  def field
-    nil
-  end
-
-  def true?(_)
-    true
-  end
-
-  def is_missing?(_)
-    false
-  end
-end
