+++ lib/scoruby/predicates/false_predicate.rb
@@ -1,0 +1,17 @@
+module Scoruby
+  module Predicates
+    class FalsePredicate
+      def field
+        nil
+      end
+
+      def true?(_)
+        false
+      end
+
+      def is_missing?(_)
+        false
+      end
+    end
+  end
+end
