+++ lib/scoruby/predicates/true_predicate.rb
@@ -1,0 +1,17 @@
+module Scoruby
+  module Predicates
+    class TruePredicate
+      def field
+        nil
+      end
+
+      def true?(_)
+        true
+      end
+
+      def is_missing?(_)
+        false
+      end
+    end
+  end
+end
