+++ spec/scoruby/predicates/simple_predicate_spec.rb
@@ -1,11 +1,11 @@
 require 'spec_helper'
 
-describe SimplePredicate do
+describe Scoruby::Predicates::SimplePredicate do
 
   context 'less or equal predicate' do
     let (:pred_string) { '<SimplePredicate field="f33" operator="lessOrEqual" value="18.8513846048894"/>' }
     let (:pred_xml) { Nokogiri::XML(pred_string); }
-    let (:predicate) { SimplePredicate.new(pred_xml.children.first) }
+    let (:predicate) { Scoruby::Predicates::SimplePredicate.new(pred_xml.children.first) }
 
     it 'returns true' do
       expect(predicate.true?(f33: 18)).to eq true
@@ -19,7 +19,7 @@ describe SimplePredicate do
   context 'is missing predicate' do
     let (:pred_string) { '<SimplePredicate field="f33" operator="isMissing"/>' }
     let (:pred_xml) { Nokogiri::XML(pred_string); }
-    let (:predicate) { SimplePredicate.new(pred_xml.children.first) }
+    let (:predicate) { Scoruby::Predicates::SimplePredicate.new(pred_xml.children.first) }
 
     it 'returns true' do
       expect(predicate.true?(f33: nil)).to eq true
@@ -37,7 +37,7 @@ describe SimplePredicate do
   context 'equals predicate' do
     let (:pred_string) { '<SimplePredicate field="f33" operator="equal" value="f2v3"/>' }
     let (:pred_xml) { Nokogiri::XML(pred_string); }
-    let (:predicate) { SimplePredicate.new(pred_xml.children.first) }
+    let (:predicate) { Scoruby::Predicates::SimplePredicate.new(pred_xml.children.first) }
 
     it 'returns true' do
       expect(predicate.true?(f33: 'f2v3')).to eq true
