+++ spec/scoruby/predicates/simple_set_predicate_spec.rb
@@ -1,10 +1,10 @@
 require 'spec_helper'
 
-describe SimpleSetPredicate do
+describe Scoruby::Predicates::SimpleSetPredicate do
 
   let (:pred_xml) { Scoruby.xml_from_string(pred_string) }
   let (:relevant_pred_xml) {  pred_xml.children[0] }
-  let (:categorical_predicate) { SimpleSetPredicate.new(relevant_pred_xml) }
+  let (:categorical_predicate) { Scoruby::Predicates::SimpleSetPredicate.new(relevant_pred_xml) }
 
   context 'quotes' do
     let (:pred_string) { '<SimpleSetPredicate field="f36" booleanOperator="isIn">
